+++ src/rtls.rs
@@ -124,7 +124,10 @@ pub fn default_tls_config() -> Arc<dyn TlsConnector> {
         // risk that the rustls features are ambiguous and no process wide default crypto provider
         // has been set yet.
         let config = rustls::ClientConfig::builder_with_provider(
+            #[cfg(feature = "tls")]
             rustls::crypto::ring::default_provider().into(),
+            #[cfg(feature = "tls-aws-lc-rs")]
+            rustls::crypto::aws_lc_rs::default_provider().into(),
         )
         .with_protocol_versions(&[&rustls::version::TLS12, &rustls::version::TLS13])
         .unwrap() // Safety: the *ring* default provider always configures ciphersuites compatible w/ both TLS 1.2 & TLS 1.3